+++ CHANGES.md
@@ -98,6 +98,7 @@ Changes in 12.5
 - CSP: use `<details><summary>` CSS deprecate .switchMenu in stylesheet.css
 - CSS add `<pre>` styling in colors.css & stylesheet.css
 - CSS FlatSIS theme Prevent body from scrolling when submenu open + scroll submenu in stylesheet.css
+- CSP use PHP instead of JS for salute header in Portal.php
 
 Changes in 12.4.3
 -----------------